refactor: direct Debug/Warn calls; split grant tests per resource (CXH-2285)
Replace the bound-log-level-in-a-variable pattern in the folder/document
collaborator pre-check with a plain if/else calling l.Debug/l.Warn directly,
matching the rest of the package.

Split grant_idempotency_test.go into folder_test.go and document_test.go
per the resource-per-file convention, with shared mock server and helpers
in collaborator_test.go.
